feat: add security to chore types API (#4)

+++ src/main/java/es/sralloza/choremanagementbot/controllers/ChoreTypesController.java
@@ -1,6 +1,7 @@
 package es.sralloza.choremanagementbot.controllers;
 
 import es.sralloza.choremanagementbot.models.custom.ChoreType;
+import es.sralloza.choremanagementbot.security.SimpleSecurity;
 import es.sralloza.choremanagementbot.services.ChoreTypesService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
+import javax.validation.constraints.NotNull;
 import java.util.List;
 
 import static org.springframework.http.HttpStatus.NO_CONTENT;
@@ -22,6 +24,8 @@ import static org.springframework.http.HttpStatus.NO_CONTENT;
 public class ChoreTypesController {
     @Autowired
     private ChoreTypesService service;
+    @Autowired
+    private SimpleSecurity security;
 
     @GetMapping()
     public List<ChoreType> listChoreTypes() {
@@ -34,13 +38,15 @@ public class ChoreTypesController {
     }
 
     @PostMapping()
-    public ChoreType createChoreType(@RequestBody @Valid ChoreType choreType) {
+    public ChoreType createChoreType(@RequestBody @NotNull @Valid ChoreType choreType) {
+        security.requireAdmin();
         return service.createChoreType(choreType);
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(value = NO_CONTENT)
     public void deleteChoreType(@PathVariable("id") String id) {
+        security.requireAdmin();
         service.deleteChoreType(id);
     }
 }
